Cliente, TipoCliente, Telefones e Enderecos

+++ src/main/java/com/lojanelioalves/api/entities/Cidade.java
@@ -10,7 +10,7 @@ public class Cidade implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nome;
 
     @ManyToOne
@@ -20,17 +20,17 @@ public class Cidade implements Serializable {
     public Cidade() {
     }
 
-    public Cidade(Integer id, String nome, Estado estado) {
+    public Cidade(Long id, String nome, Estado estado) {
         this.id = id;
         this.nome = nome;
         this.estado = estado;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
